Archived function that stores the day's tide
